Reorganize lib to fit distribution conventions.
