+++ .gitignore
@@ -2,4 +2,5 @@
 build
 dist
 MANIFEST
+__pycache__
