+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include *.py AUTHORS README.rst *.h MANIFEST.in LICENSE
+include *.py AUTHORS README.rst snappy/*.h MANIFEST.in LICENSE
