+++ snappy/__init__.py
@@ -1,0 +1,11 @@
+from snappy import (
+	compress,
+	decompress,
+	uncompress,
+	stream_compress,
+	stream_decompress,
+	StreamCompressor,
+	StreamDecompressor,
+	UncompressError,
+	isValidCompressed,
+)
