+++ snappy/snappymodule.cc
